# Remove ReferenceQueue from TaintedMap implementation and purge inline