--- a/dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapEmptyBenchmark.java
+++ b/dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapEmptyBenchmark.java
@@ -3,6 +3,8 @@ package com.datadog.iast.taint;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
+import com.datadog.iast.taint.TaintedMap.WithPurgeInline;
+import com.datadog.iast.taint.TaintedMap.WithPurgeQueue;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -15,6 +17,7 @@ import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
 import org.openjdk.jmh.annotations.Warmup;
+import org.openjdk.jmh.infra.BenchmarkParams;
 import org.openjdk.jmh.infra.Blackhole;
 
 @Warmup(iterations = 2, time = 1000, timeUnit = MILLISECONDS)
@@ -27,16 +30,39 @@ public class TaintedMapEmptyBenchmark {
 
   private static final int OP_COUNT = 1024;
   private TaintedMap map;
+
   private final Object anyObject = new Object();
 
   @Setup(Level.Iteration)
-  public void setup() {
-    map = new TaintedMap();
+  public void setup(BenchmarkParams params) {
+    if (params.getBenchmark().endsWith("purgeQueue")) {
+      map = new WithPurgeQueue();
+    } else if (params.getBenchmark().endsWith("purgeInline")) {
+      map = new WithPurgeInline();
+    } else {
+      map = TaintedMap.NoOp.INSTANCE;
+    }
   }
 
   @Benchmark
   @OperationsPerInvocation(OP_COUNT)
-  public void getFromEmptyMap(final Blackhole bh) {
+  public void baseline(final Blackhole bh) {
+    for (int i = 0; i < OP_COUNT; i++) {
+      bh.consume(map.get(anyObject));
+    }
+  }
+
+  @Benchmark
+  @OperationsPerInvocation(OP_COUNT)
+  public void purgeQueue(final Blackhole bh) {
+    for (int i = 0; i < OP_COUNT; i++) {
+      bh.consume(map.get(anyObject));
+    }
+  }
+
+  @Benchmark
+  @OperationsPerInvocation(OP_COUNT)
+  public void purgeInline(final Blackhole bh) {
     for (int i = 0; i < OP_COUNT; i++) {
       bh.consume(map.get(anyObject));
     }