--- a/dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapGetsBenchmark.java
+++ b/dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapGetsBenchmark.java
@@ -1,9 +1,12 @@
 package com.datadog.iast.taint;
 
+import static com.datadog.iast.taint.TaintedMap.WithPurgeQueue.DEFAULT_FLAT_MODE_THRESHOLD;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import com.datadog.iast.model.Range;
+import com.datadog.iast.taint.TaintedMap.WithPurgeInline;
+import com.datadog.iast.taint.TaintedMap.WithPurgeQueue;
 import java.util.ArrayList;
 import java.util.List;
 import org.openjdk.jmh.annotations.Benchmark;
@@ -17,7 +20,9 @@ import org.openjdk.jmh.annotations.OutputTimeUnit;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
+import org.openjdk.jmh.annotations.TearDown;
 import org.openjdk.jmh.annotations.Warmup;
+import org.openjdk.jmh.infra.BenchmarkParams;
 import org.openjdk.jmh.infra.Blackhole;
 
 @Warmup(iterations = 1, time = 1000, timeUnit = MILLISECONDS)
@@ -28,7 +33,7 @@ import org.openjdk.jmh.infra.Blackhole;
 @State(Scope.Benchmark)
 public class TaintedMapGetsBenchmark {
 
-  private static final int INITIAL_OP_COUNT = TaintedMap.DEFAULT_FLAT_MODE_THRESHOLD;
+  private static final int INITIAL_OP_COUNT = DEFAULT_FLAT_MODE_THRESHOLD >> 1;
   private static final int OP_COUNT = 1024;
 
   private TaintedMap map;
@@ -36,8 +41,14 @@ public class TaintedMapGetsBenchmark {
   private List<Object> initialObjectList;
 
   @Setup(Level.Iteration)
-  public void setup() {
-    map = new TaintedMap();
+  public void setup(BenchmarkParams params) {
+    if (params.getBenchmark().endsWith("purgeQueue")) {
+      map = new WithPurgeQueue();
+    } else if (params.getBenchmark().endsWith("purgeInline")) {
+      map = new WithPurgeInline();
+    } else {
+      map = TaintedMap.NoOp.INSTANCE;
+    }
     initialObjectList = new ArrayList<>(INITIAL_OP_COUNT);
     objectList = new ArrayList<>(OP_COUNT);
     for (int i = 0; i < INITIAL_OP_COUNT; i++) {
@@ -52,17 +63,32 @@ public class TaintedMapGetsBenchmark {
     }
   }
 
-  @Benchmark
-  @OperationsPerInvocation(OP_COUNT)
-  public void getsBaseline(final Blackhole bh) {
-    for (int i = 0; i < OP_COUNT; i++) {
-      bh.consume(objectList.get(i));
+  @TearDown(Level.Iteration)
+  public void tearDown() {
+    if (map.isFlat()) {
+      throw new IllegalStateException("Map should never go flat during the test");
     }
   }
 
   @Benchmark
   @OperationsPerInvocation(OP_COUNT)
-  public void gets(final Blackhole bh) {
+  public void baseline(final Blackhole bh) {
+    for (int i = 0; i < OP_COUNT; i++) {
+      bh.consume(map.get(objectList.get(i)));
+    }
+  }
+
+  @Benchmark
+  @OperationsPerInvocation(OP_COUNT)
+  public void purgeQueue(final Blackhole bh) {
+    for (int i = 0; i < OP_COUNT; i++) {
+      bh.consume(map.get(objectList.get(i)));
+    }
+  }
+
+  @Benchmark
+  @OperationsPerInvocation(OP_COUNT)
+  public void purgeInline(final Blackhole bh) {
     for (int i = 0; i < OP_COUNT; i++) {
       bh.consume(map.get(objectList.get(i)));
     }