--- a/dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapPutsBenchmark.java
+++ b/dd-java-agent/agent-iast/src/jmh/java/com/datadog/iast/taint/TaintedMapPutsBenchmark.java
@@ -4,8 +4,6 @@ import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 
 import com.datadog.iast.model.Range;
-import com.datadog.iast.taint.TaintedMap.WithPurgeInline;
-import com.datadog.iast.taint.TaintedMap.WithPurgeQueue;
 import datadog.trace.test.util.CircularBuffer;
 import java.util.ArrayList;
 import java.util.IdentityHashMap;
@@ -22,11 +20,9 @@ import org.openjdk.jmh.annotations.OutputTimeUnit;
 import org.openjdk.jmh.annotations.Scope;
 import org.openjdk.jmh.annotations.Setup;
 import org.openjdk.jmh.annotations.State;
-import org.openjdk.jmh.annotations.TearDown;
 import org.openjdk.jmh.annotations.Timeout;
 import org.openjdk.jmh.annotations.Warmup;
 import org.openjdk.jmh.infra.BenchmarkParams;
-import org.openjdk.jmh.infra.Blackhole;
 
 @Warmup(iterations = 3, time = 1000, timeUnit = MILLISECONDS)
 @Measurement(iterations = 5, time = 1000, timeUnit = MILLISECONDS)
@@ -37,7 +33,7 @@ import org.openjdk.jmh.infra.Blackhole;
 @State(Scope.Benchmark)
 public class TaintedMapPutsBenchmark {
 
-  private static final int INITIAL_OP_COUNT = WithPurgeQueue.DEFAULT_FLAT_MODE_THRESHOLD >> 1;
+  private static final int INITIAL_OP_COUNT = 1 << 12;
   private static final int OP_COUNT = 1024;
 
   private static final Range[] EMPTY_RANGES = new Range[0];
@@ -48,35 +44,23 @@ public class TaintedMapPutsBenchmark {
 
   @Setup(Level.Iteration)
   public void setup(BenchmarkParams params) {
+    final boolean baseline = params.getBenchmark().endsWith("baseline");
+    map = baseline ? TaintedMap.NoOp.INSTANCE : new TaintedMap.TaintedMapImpl();
     gcHandler = new GarbageCollectorHandler(OP_COUNT);
-    if (params.getBenchmark().endsWith("purgeQueue")) {
-      map = new WithPurgeQueue();
-    } else if (params.getBenchmark().endsWith("purgeInline")) {
-      map = new WithPurgeInline();
-    } else {
-      map = TaintedMap.NoOp.INSTANCE;
-    }
     initialObjectList = new ArrayList<>(INITIAL_OP_COUNT);
     for (int i = 0; i < INITIAL_OP_COUNT; i++) {
       final Object k = new Object();
       initialObjectList.add(k);
-      map.put(new TaintedObject(k, EMPTY_RANGES, map.getReferenceQueue()));
-    }
-  }
-
-  @TearDown(Level.Iteration)
-  public void tearDown() {
-    if (map.isFlat()) {
-      throw new IllegalStateException("Map should never go flat during the test");
+      map.put(new TaintedObject(k, EMPTY_RANGES));
     }
   }
 
   @Benchmark
   @OperationsPerInvocation(OP_COUNT)
-  public void baseline(final Blackhole bh) {
+  public void baseline() {
     for (int i = 0; i < OP_COUNT; i++) {
       final Object k = new Object();
-      final TaintedObject to = new TaintedObject(k, EMPTY_RANGES, map.getReferenceQueue());
+      final TaintedObject to = new TaintedObject(k, EMPTY_RANGES);
       gcHandler.add(to);
       map.put(to);
     }
@@ -84,21 +68,10 @@ public class TaintedMapPutsBenchmark {
 
   @Benchmark
   @OperationsPerInvocation(OP_COUNT)
-  public void purgeQueue() {
+  public void puts() {
     for (int i = 0; i < OP_COUNT; i++) {
       final Object k = new Object();
-      final TaintedObject to = new TaintedObject(k, EMPTY_RANGES, map.getReferenceQueue());
-      gcHandler.add(to);
-      map.put(to);
-    }
-  }
-
-  @Benchmark
-  @OperationsPerInvocation(OP_COUNT)
-  public void purgeInline() {
-    for (int i = 0; i < OP_COUNT; i++) {
-      final Object k = new Object();
-      final TaintedObject to = new TaintedObject(k, EMPTY_RANGES, map.getReferenceQueue());
+      final TaintedObject to = new TaintedObject(k, EMPTY_RANGES);
       gcHandler.add(to);
       map.put(to);
     }