--- a/dd-java-agent/agent-iast/src/main/java/com/datadog/iast/taint/TaintedObject.java
+++ b/dd-java-agent/agent-iast/src/main/java/com/datadog/iast/taint/TaintedObject.java
@@ -4,7 +4,6 @@ import static com.datadog.iast.taint.TaintedMap.POSITIVE_MASK;
 
 import com.datadog.iast.model.Range;
 import datadog.trace.api.Config;
-import java.lang.ref.ReferenceQueue;
 import java.lang.ref.WeakReference;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -17,11 +16,8 @@ public class TaintedObject extends WeakReference<Object> {
   @Nullable TaintedObject next;
   private Range[] ranges;
 
-  public TaintedObject(
-      final @Nonnull Object obj,
-      final @Nonnull Range[] ranges,
-      final @Nullable ReferenceQueue<Object> queue) {
-    super(obj, queue);
+  public TaintedObject(final @Nonnull Object obj, final @Nonnull Range[] ranges) {
+    super(obj);
     this.positiveHashCode = System.identityHashCode(obj) & POSITIVE_MASK;
     // ensure ranges never go over the limit
     if (ranges.length > MAX_RANGE_COUNT) {