--- a/dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/propagation/PropagationModuleTest.groovy
+++ b/dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/propagation/PropagationModuleTest.groovy
@@ -471,7 +471,7 @@ class PropagationModuleTest extends IastModuleImplTestBase {
   private TaintedObject getTaintedObject(final Object target) {
     if (target instanceof Taintable) {
       final source = (target as Taintable).$$DD$getSource() as Source
-      return source == null ? null : new TaintedObject(target, Ranges.forObject(source), null)
+      return source == null ? null : new TaintedObject(target, Ranges.forObject(source))
     }
     return ctx.getTaintedObjects().get(target)
   }