--- a/dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/taint/TaintedObjectsLogTest.groovy
+++ b/dd-java-agent/agent-iast/src/test/groovy/com/datadog/iast/taint/TaintedObjectsLogTest.groovy
@@ -76,7 +76,6 @@ class TaintedObjectsLogTest extends DDSpecification {
     then:
     taintedObjects.size() == 1
     taintedObjects.iterator().size() == 1
-    !taintedObjects.flat
   }
 }