# What Does This Do
Replaces the original tainted map implementation based on a queue by a map that purges stale entries while reading or writing data.

# Motivation
We want to move towards a global instance of the tainted map (this way we wont require access to the current tracing context) in order to interact with the map. Previous implementation might introduce delays (when the map is going to be purged) that makes it not very suitable for a global instance.

# Additional Notes

Jira ticket: [PROJ-IDENT]

Some benchmarks

**Before**

| Benchmark | Mode | Cnt | Score | Error | Units |
| --------------- | -------- | ---- | -------- | ------- | ------- |
| TaintedMapEmptyBenchmark.getFromEmptyMap  | avgt  |  9  | 2.628 | ± 0.096 | ns/op |
| TaintedMapGetsBenchmark.gets | avgt | 9 | 4.485 | ± 0.949 | ns/op |
| TaintedMapGetsBenchmark.getsBaseline | avgt | 9 | 2.586 |± 0.091 |  ns/op |
| TaintedMapPutsBenchmark.puts | avgt | 9  | 56.547 | ± 2.568 | ns/op |
| TaintedMapPutsBenchmark.putsBaseline | avgt | 9 | 54.387 | ± 2.855 | ns/op |

**After**

| Benchmark | Mode | Cnt | Score | Error | Units |
| --------------- | -------- | ---- | -------- | ------- | ------- |
| TaintedMapEmptyBenchmark.baseline |     avgt   |  9   |  2.987 | ±  0.180 |  ns/op | 
| TaintedMapEmptyBenchmark.purgeInline |  avgt  |   9  |   3.320 | ±  0.157  | ns/op | 
| TaintedMapEmptyBenchmark.purgeQueue |   avgt |    9 |    2.872 |  ±  0.123 |  ns/op | 
| TaintedMapGetsBenchmark.baseline  |     avgt |    9  |   2.649 |  ±  0.264 |  ns/op | 
| TaintedMapGetsBenchmark.purgeInline  |  avgt  |   9 |    6.902 |  ±  0.370 |  ns/op | 
| TaintedMapGetsBenchmark.purgeQueue  |   avgt   |  9  |   5.036 | ±  0.555 |  ns/op | 
| TaintedMapPutsBenchmark.baseline   |    avgt |   15  |  90.647 | ±  3.215 |  ns/op | 
| TaintedMapPutsBenchmark.purgeInline |   avgt  |  15  | 116.038 | ±  3.248 |  ns/op | 
| TaintedMapPutsBenchmark.purgeQueue   |  avgt  |  15  | 184.431 | ± 10.006 |  ns/op | 

<!--
# Opening vs Drafting a PR:
When opening a pull request, please open it as a draft to not auto assign reviewers before you feel the pull request is in a reviewable state.

# Linking a JIRA ticket:
Please link your JIRA ticket by adding its identifier between brackets (ex [PROJ-IDENT]) in the PR description, not the title.
This requirement only applies to Datadog employees.
-->